feat: API 호출해서 송금액 계산하는 기능과 테스트

+++ src/main/java/com/jaehee/currencyconverter/service/CurrencyService.java
@@ -15,7 +15,7 @@ public class CurrencyService {
     }
 
     //환율, 금액으로 송금액 계산하기
-    public Double getRemittance(Double exchangeRate, int amount){
+    public Double getRemittance(Double exchangeRate, Double amount){
         return exchangeRate*amount;
     };
 }
